src/offensivetextdetectionservice/requirements.txt
kafka-python==1.4.6
alt-profanity-check